refactor: change the name of the main algorithm, replace List with ArrayList, change LayoutUtils to object, move the adjustSpeedAndApplyForces method to LayoutUtils
