+++ src/main/kotlin/algorithms/layout/Region.kt
@@ -1,17 +1,15 @@
 package algorithms.layout
 
-import kotlin.math.max
-import kotlin.math.min
 import kotlin.math.pow
 import kotlin.math.sqrt
 
-class Region(private val vertices: List<Vertex>) {
-    private val utils = FA2Utils()
+class Region(private val vertices: ArrayList<Vertex>) {
+    private val utils = LayoutUtils
     var mass = 0.0
     var massCenterX = 0.0
     var massCenterY = 0.0
     private var size = 0.0
-    private var subregions: MutableList<Region> = mutableListOf()
+    private var subregions = ArrayList<Region>()
 
     init {
         updateMassAndGeometry()
@@ -40,10 +38,10 @@ class Region(private val vertices: List<Vertex>) {
 
     fun buildSubRegions() {
         if (vertices.size > 1) {
-            val topLeftVertices = mutableListOf<Vertex>()
-            val bottomLeftVertices = mutableListOf<Vertex>()
-            val topRightVertices = mutableListOf<Vertex>()
-            val bottomRightVertices = mutableListOf<Vertex>()
+            val topLeftVertices = arrayListOf<Vertex>()
+            val bottomLeftVertices = arrayListOf<Vertex>()
+            val topRightVertices = arrayListOf<Vertex>()
+            val bottomRightVertices = arrayListOf<Vertex>()
 
             for (vertex in vertices) {
                 if (vertex.x < massCenterX) {
@@ -61,13 +59,13 @@ class Region(private val vertices: List<Vertex>) {
                 }
             }
 
-            listOf(topLeftVertices, bottomLeftVertices, topRightVertices, bottomRightVertices).forEach { regionVertices ->
+            arrayListOf(topLeftVertices, bottomLeftVertices, topRightVertices, bottomRightVertices).forEach { regionVertices ->
                 if (regionVertices.isNotEmpty()) {
                     if (regionVertices.size < vertices.size) {
                         subregions.add(Region(regionVertices))
                     } else {
                         regionVertices.forEach {
-                            subregions.add(Region(listOf(it)))
+                            subregions.add(Region(arrayListOf(it)))
                         }
                     }
                 }
@@ -94,76 +92,10 @@ class Region(private val vertices: List<Vertex>) {
     }
 
     fun applyForceOnNodes(
-        vertices: List<Vertex>,
+        vertices: ArrayList<Vertex>,
         theta: Double,
         coefficient: Double = 0.0,
     ) {
         vertices.forEach { applyForce(it, theta, coefficient) }
     }
-}
-
-fun adjustSpeedAndApplyForces(
-    vertices: List<Vertex>,
-    speed: Double,
-    speedEfficiency: Double,
-    jitterTolerance: Double,
-): Pair<Double, Double> {
-    var totalSwinging = 0.0
-    var totalEffectiveTraction = 0.0
-    vertices.forEach { vertex ->
-        val swinging = sqrt((vertex.oldDx - vertex.dx).pow(2) + (vertex.oldDy - vertex.dy).pow(2))
-        totalSwinging += vertex.mass * swinging
-        val effectiveTraction = 0.5 * sqrt((vertex.oldDx + vertex.dx).pow(2) + (vertex.oldDy + vertex.dy).pow(2))
-        totalEffectiveTraction += vertex.mass * effectiveTraction
-    }
-    val estimatedOptimalJitterTolerance = .05 * sqrt(vertices.size.toDouble())
-    val minJT = sqrt(estimatedOptimalJitterTolerance)
-    val maxJT = sqrt(estimatedOptimalJitterTolerance)
-
-    val jt =
-        jitterTolerance *
-            max(
-                minJT,
-                min(
-                    maxJT,
-                    estimatedOptimalJitterTolerance * totalEffectiveTraction / (
-                        vertices.size * vertices.size
-                    ),
-                ),
-            )
-    val minSpeedEfficiency = 0.05
-
-    var newSpeedEfficiency = speedEfficiency
-    if (totalEffectiveTraction > 2.0 && totalSwinging / totalEffectiveTraction > 2.0) {
-        if (newSpeedEfficiency > minSpeedEfficiency) {
-            newSpeedEfficiency *= .5
-        }
-    }
-
-    val targetSpeed =
-        if (totalSwinging == 0.0) {
-            Double.MAX_VALUE
-        } else {
-            jt * newSpeedEfficiency * totalEffectiveTraction / totalSwinging
-        }
-
-    if (totalSwinging > jt * totalEffectiveTraction) {
-        if (newSpeedEfficiency > minSpeedEfficiency) {
-            newSpeedEfficiency *= .7
-        }
-    } else if (speed < 1000) {
-        newSpeedEfficiency *= 1.3
-    }
-
-    val maxRise = .5
-    val newSpeed = speed + min(targetSpeed - speed, maxRise * speed)
-
-    vertices.forEach {
-        val swinging = it.mass * sqrt((it.oldDx - it.dx).pow(2) + (it.oldDy - it.dy).pow(2))
-        val factor = speed / (1.0 + sqrt(speed * swinging))
-        it.x += (it.dx * factor)
-        it.y += (it.dy * factor)
-    }
-
-    return newSpeed to minSpeedEfficiency
 }
